Moving some tests that Go can also hit to check we match Go
