+++ compiler/natives/src/slices/slices_test.go
@@ -1,0 +1,9 @@
+//go:build js
+
+package slices
+
+import "testing"
+
+func TestGrow(t *testing.T) {
+	t.Skip("testing.AllocsPerRun not supported in GopherJS")
+}
